Fixed piece movement module logic

+++ src/piece_movement/king_movement.rs
@@ -1,3 +1,3 @@
-pub fn get_king_moves(occupied_white: Vec<(i8, i8)>, occupied_black: Vec<(i8, i8)>) {
+pub fn get_king_moves(occupied_white: Vec<(i32, i32)>, occupied_black: Vec<(i32, i32)>) {
     println!("KING!");
 }
